pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cesure"
version = "0.1.0"
description = "Art Gallery Website"
readme = "README.md"
requires-python = ">=3.13"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "sqlalchemy>=2.0.0",
    "alembic>=1.7.0",
    "psycopg2-binary>=2.9.0",
    "fastapi>=0.103.0",
    "uvicorn>=0.23.0",
    "pydantic>=2.4.0",
    "python-jose>=3.3.0",
    "passlib>=1.7.4",
    "bcrypt>=4.0.1",
    "python-multipart>=0.0.5",
    "httpx>=0.24.0",
    "pydantic[email]>=2.4.0",
    "jinja2>=3.1.6",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "isort",
    "flake8",
    "debugpy",
]